+++ CHANGES.rst
@@ -28,6 +28,7 @@ Fixes:
 - Fix ``celery>4.2`` not found because of application path modification.
 - Fix invalid handling of ``wps_processes.yml`` reference in ``weaver.ini`` when specified as relative path to
   configuration directory.
+- Fix handling of ``WPS<->CWL`` I/O merge of ``data_format`` field against ``supported_formats`` with ``pywps>=4.2.4``.
 
 `1.7.0 <https://github.com/crim-ca/weaver/tree/1.7.0>`_ (2020-05-15)
 ========================================================================